reactivated animation of legacy IE<=8
Although there are good reasons to deactivate the animation
there is a strong client demand to have animation in place
on legacy IE <= 8.

Still, we need to improve and look into the problem with
animation on IE8 if the div layer is hidden.

deactivating the anination was introduced as part of the commit
3ec0d49fd2192b2109259045202edc78c60d3d8d